+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Erl2ex is currently pre-alpha software. Expect significant backwards-incompatibl
 ## v0.0.7 (not yet released)
 
 *   Support invoking a constant macro as a function name.
+*   If a function's argument pattern looks like a keyword block, it tried to codegen as such. Fixed.
 
 ## v0.0.6 (2016-01-19)